# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Holiday Plan API",
        "description": "API for managing holidays",
        "contact": {
            "name": "Support Team",
            "email": "support@holidayapi.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/holidays": {
            "get": {
                "tags": [
                    "Holiday Plans"
                ],
                "summary": "Get all holiday plans",
                "operationId": "933c70da048ccd8c011f95ffda01512e",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HolidayPlanResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Holiday Plans"
                ],
                "summary": "Create a new holiday plan",
                "operationId": "0953d7c01b787efe0411db7ae3dea898",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HolidayPlan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HolidayPlan"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/holidays/{id}": {
            "get": {
                "tags": [
                    "Holiday Plans"
                ],
                "summary": "Get a holiday plan",
                "operationId": "eaa37884babc4d705389523406c3c873",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the holiday plan",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HolidayPlanResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Holiday plan not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "message": "Holiday plan not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "message": "Unauthorized"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Holiday Plans"
                ],
                "summary": "Update a holiday plan",
                "operationId": "84b52c78ba10a6b4e9d1deae9bd93e13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the holiday plan",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HolidayPlan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HolidayPlanResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Holiday plan not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "message": "Holiday plan not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "message": "Unauthorized"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Holiday Plans"
                ],
                "summary": "Delete a holiday plan",
                "operationId": "76185d4497e4ea02ad7f9afe8a00ec1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the holiday plan",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "message": "Holiday plan deleted successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Holiday plan not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "message": "Holiday plan not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "message": "Unauthorized"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/holiday/{id}/pdf": {
            "get": {
                "tags": [
                    "Holiday Plans"
                ],
                "summary": "Get a holiday plan in PDF",
                "operationId": "a53ba83f0e66c3e4254bc008ade7d6b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the holiday plan",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/pdf": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/holiday/pdf": {
            "get": {
                "tags": [
                    "Holiday Plans"
                ],
                "summary": "Get all holiday plans in PDF",
                "operationId": "f5bc840eef04adb0d091b71ccb459deb",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/pdf": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "HolidayPlan": {
                "required": [
                    "title",
                    "description",
                    "date_from",
                    "date_to",
                    "location"
                ],
                "properties": {
                    "title": {
                        "description": "The title of the holiday plan",
                        "type": "string",
                        "example": "Viagem Cultural a Europa"
                    },
                    "description": {
                        "description": "The description of the holiday plan",
                        "type": "string",
                        "example": "Tour cultural por museus e galerias de arte em várias cidades da Europa."
                    },
                    "date_from": {
                        "description": "The start date of the holiday plan",
                        "type": "string",
                        "format": "date"
                    },
                    "date_to": {
                        "description": "The end date of the holiday plan",
                        "type": "string",
                        "format": "date"
                    },
                    "location": {
                        "description": "The location of the holiday plan",
                        "type": "string",
                        "example": "Paris"
                    }
                },
                "type": "object"
            },
            "HolidayPlanResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/HolidayPlan"
                    },
                    {
                        "properties": {
                            "id": {
                                "description": "The ID of the user who created the holiday plan",
                                "type": "integer",
                                "example": 1
                            },
                            "status": {
                                "description": "The status of the holiday plan",
                                "type": "string",
                                "example": "pending"
                            },
                            "user_id": {
                                "description": "The ID of the user who created the holiday plan",
                                "type": "integer",
                                "example": 10
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação necessária. Use 'Bearer {token}' no campo de autorização.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}